requirements.txt
accelerate==0.28.0
audio-separator==0.17.2
av==12.1.0
bitsandbytes==0.43.1
decord==0.6.0
diffusers
einops==0.8.0
insightface==0.7.3
librosa==0.10.2.post1
mediapipe[vision]==0.10.14
mlflow==2.13.1
moviepy==1.0.3
numpy==1.26.4
omegaconf==2.3.0
onnx2torch==1.5.14
onnx==1.16.1
onnxruntime-gpu==1.18.0
opencv-python==4.10.0.84
pillow==10.3.0
pyyaml==6.0.1
setuptools==70.0.0
transformers==4.39.2
xformers==0.0.25.post1
isort==5.13.2
pre-commit==3.7.1
scipy==1.13.1
imageio==2.34.2
lmdb==1.4.1
rich==13.7.1
ffmpeg-python==0.2.0
scikit-image==0.24.0
albumentations==1.4.10
matplotlib==3.9.0
imageio-ffmpeg==0.5.1
tyro==0.8.5
gradio==5.1.0
pykalman==0.9.7
tensorboardX==2.6.2.2
pydub